### 2021-04-05 #3149 @whedon
[PRE REVIEW]: datasailr - An R Package for Row by Row Data Processing, Using DataSailr Script
**Submitting author:** @niceume (<a href="http://orcid.org/0000-0001-7015-621X">Toshihiro Umehara</a>)
**Repository:** <a href="https://github.com/niceume/datasailr" target ="_blank">https://github.com/niceume/datasailr</a>
**Version:** 0.8.7
**Reviewers:** @DrMattG and @crvernon
