+++ inst/unit_tests/test_if.R
@@ -3,26 +3,25 @@ test_if <- function(){
   mtcars[,"name"] = rownames(mtcars)
 
   code = '
-power = \'\'
-if( hp > 145 ){
-  power = "powerful"
-}else if( 145 >= hp && hp > 0){
-  power = "low power"
-}else{
-  print("hp variable has missing value")
-}
+    power = ""
+    if( hp > 145 ){
+      power = "powerful"
+    }else if( 145 >= hp && hp > 0){
+      power = "low power"
+    }else{
+      print("hp variable has missing value")
+    }
 
-efficient = ""
-if( mpg > 20){
-  efficient = "efficient"
-}else if( 20 >= mpg && mpg > 0 ){
-  efficient = "inefficient"
-}else{
-  print("mpg variable has missing value")
-}
+    efficient = ""
+    if( mpg > 20){
+      efficient = "efficient"
+    }else if( 20 >= mpg && mpg > 0 ){
+      efficient = "inefficient"
+    }else{
+      print("mpg variable has missing value")
+    }
 
-print(cyl)
-description = name + " is " + power + " " + efficient + " car"
+    description = name + " is " + power + " " + efficient + " car"
   '
 
   mtcars_result = datasailr::sail(mtcars, code)
