### 2021-04-09 #3166 @whedon
[REVIEW]: datasailr - An R Package for Row by Row Data Processing, Using DataSailr Script
**Submitting author:** @niceume (<a href="http://orcid.org/0000-0001-7015-621X">Toshihiro Umehara</a>)
**Repository:** <a href="https://github.com/niceume/datasailr" target ="_blank">https://github.com/niceume/datasailr</a>
**Version:** joss0.8.8-3

@DrMattG and @crvernon, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
