### 2021-04-15 niceume/datasailr#3 @DrMattG
Add Community guidelines
(https://github.com/openjournals/joss-reviews/issues/3166)
Community guidelines: Add clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
