### 2021-04-15 niceume/datasailr#4 @DrMattG
State target audience 
(https://github.com/openjournals/joss-reviews/issues/3166)
A statement of need: Need to "...clearly state what problems the software is designed to solve and who the target audience is".

I think you could add some more to the documentation (README) to state who the target audience are. You state it nicely in the manuscript - i.e. users who are used to SAS
