### 2021-04-15 niceume/datasailr#5 @DrMattG
Consider making the function test_sail() an internal function
(https://github.com/openjournals/joss-reviews/issues/3166)
Consider making the function test_sail() an internal function. Or better explain how this function might be useful for the package user. I'm not sure how I would/should use this function whilst using your package
