### 2021-04-15 niceume/datasailr#6 @DrMattG
What is the purpose of the author() function?
(https://github.com/openjournals/joss-reviews/issues/3166)
What is the function of the function author()? I can not see how this is useful to the user (above and beyond citation("datasailr"))
