### 2021-04-15 3266d51
Update sail.Rd
Some minor typos (https://github.com/openjournals/joss-reviews/issues/3166)
