### 2021-04-15 ead5b3d
Update sail.Rd
Some minor typos (https://github.com/openjournals/joss-reviews/issues/3166)
