+++ man/sail.Rd
@@ -4,7 +4,7 @@
 \alias{sail}
 \title{Run 'DataSailr' script and manipulate dataframe }
 \description{
-  datasailr::sail() is the main function of this 'DataSailr' package. This function takes data.frame as the 1st argument and 'DataSailr' script as the 2nd argument. 
+  datasailr::sail() is the main function of this 'DataSailr' package. This function takes a data.frame as the 1st argument and a 'DataSailr' script as the 2nd argument. 
 }
 \usage{
   sail( df, code, fullData = TRUE, rowname = "_rowname_", stringsAsFactors = FALSE)
@@ -17,7 +17,7 @@
   \item{stringsAsFactors}{ Columns of character vectors (= string vectors) that appear as left-hand-side(LHS) variables in 'DataSailr' scripts can be returned as either character vectors or factor vectors. If stringsAsFactors is set to FALSE(default), character columns are returned as character vectors. If stringsAsFactors is set to TRUE, character columns are returned as factor vectors. }
 }
 \value{
-  If fullData argument is set FALSE, return value is a data.frame. If fullData is TRUE, the result is created by cbind(), which concatenate the original input and newly created dataframe. For example, if the original input is 'data.table', the return value is 'data.table'.
+  If fullData argument is set FALSE, return value is a data.frame. If fullData is TRUE, the result is created by cbind(), which concatenates the original input and newly created dataframe. For example, if the original input is 'data.table', the return value is 'data.table'.
 }
 \examples{
 library(datasailr)
