### 2021-04-15 4825742
Update sail.Rd
Some minor typos (https://github.com/openjournals/joss-reviews/issues/3166)
