### 2021-04-15 niceume/datasailr#8 @DrMattG
Expand on summary and statement of need in the manuscript
(https://github.com/openjournals/joss-reviews/issues/3166)
In the paper (https://github.com/niceume/datasailr/blob/joss_submission/paper.md) [line 6 in the summary] it would be good if you can expand on this a little with citation - I agree the data manipulation/cleaning is the most important part of data analysis. Perhaps expand on the need for row-wise (and specifically not column-wise ) manipulation (you give a good use-case in your website text). I think this will give a better "justification" for the need for the package.
