### 2021-04-15 niceume/datasailr#9 @crvernon
JOSS Review [crvernon]: GitHub actions for CI
**Review reference:** openjournals/joss-reviews#3166

Consider setting up CI for your test suite to run via GitHub actions or the like per push and/or PR on at least one OS.

See https://github.com/r-lib/actions for reference.  There are other resources available as well.
