### 2021-04-15 niceume/datasailr#10 @crvernon
JOSS Review [crvernon]:  Code formatting in paper
**Review reference:** openjournals/joss-reviews#3166

The code examples in the `DataSailr Script` section of your paper are very cluttered upon rendering.

If you wish to add code in this section, please do so under a header using R formatting like the following:

```
# DataSailr Script
...

## Example

```r
library(datasailr)
data(mtcars)

result = datasailr::(mtcars, '
// Comments...
if (hp > 100) {
    ...
}')
```

Which will render (see following comment on R formatting):

# DataSailr Script
...

## Example

```r
library(datasailr)
data(mtcars)

result = datasailr::(mtcars, '
// Comments...
if (hp > 100) {
    ...
}')
```

## Output

...

The R formatting will conduct the appropriate highlighting upon compile which you can test here:  https://whedon.theoj.org/ or by running `@whedon generate pdf from branch joss_submission` in our review issue once you push the changes.
