+++ man/author.Rd
@@ -1,12 +1,0 @@
-%File author.Rd
-
-\name{author}
-\alias{author}
-\title{show author of this package }
-\description{
-  datasailr::author() shows author information of this package.
-}
-\usage{
-  author()
-}
-
