### 2021-04-17 5434a52
test_sail() is now an internal function.
