+++ man/test_sail.Rd
@@ -1,19 +1,0 @@
-%File  test_sail.Rd
-
-\name{test_sail}
-\alias{test_sail}
-\title{run DataSailr test scripts }
-\description{
-  datasailr::test_sail() run test scripts for 'DataSailr' package.
-}
-\usage{
-  test_sail()
-}
-\arguments{
-  No arguments required.
-}
-\examples{
-library(datasailr)
-test_sail()
-}
-
