+++ NAMESPACE
@@ -1,3 +1,3 @@
 useDynLib(datasailr, .registration=TRUE)
-exportPattern("^[[:alpha:]]+")
+export("sail")
 importFrom(Rcpp, evalCpp)
