### 2021-04-17 niceume/datasailr#3 @niceume
Community guidelines are now added in 962cea017fe88329361930bd5da57fdbb8daf3db .
