### 2021-04-17 niceume/datasailr#5 @niceume
test_sail() function is now an internal function. 5434a529227272cf2e4b8f64e5081a48ac9cd00c
