### 2021-04-17 niceume/datasailr#6 @niceume
There is no meaning for author() function... The function and its documentation are deleted. c9bddaf49a4641dd291dccdf194a5623d9589ba6
