+++ paper.bib
@@ -1,3 +1,19 @@
+@book{pyle1999,
+  title     = {Data Preparation for Data Mining},
+  author    = {Dorian Pyle},
+  year      = {1999},
+  publisher = {Morgan Kaufmann},
+  address   = {Los Altos, California},
+}
+
+@book{garcia2015,
+  title     = {Data Preprocessing in Data Mining},
+  author = {Salvador García and Julián Luengo and Francisco Herrera},
+  year      = {2015},
+  publisher = {Springer International Publishing},
+  address   = {Switzerland}
+}
+
 @Manual{dplyr:2021,
   title = {dplyr: A Grammar of Data Manipulation},
   author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller},
