### 2021-04-17 niceume/datasailr#8 @niceume
Thank you very much for your suggestions.
Citations are added. fe3d7bb794d66f171cdd70082d974452953c6623
Justification using example is added. 5763b77d829c5e6bd0f1100f61db8c6b9c730074
