### 2021-04-17 niceume/datasailr#9 @niceume
Thank you very much for your suggestion!

To build this package,  additional tools such as autotools, bison and flex are required. Just using the r-lib/actions, build does not succeed. At now, I do not have  enough knowledge to set up additional tools on Github actions.

This package is accepted by CRAN, and it shows this package is successfully built on multiple platforms.  https://cran.r-project.org/web/checks/check_results_datasailr.html
