### 2021-04-19 niceume/datasailr#9 @crvernon
This approach seems reasonable for now.  Thanks!
