### 2021-04-19 niceume/datasailr#12 @crvernon
JOSS Review [crvernon]:  Readme code example incorrect
**Review reference:** openjournals/joss-reviews#3166

The example in your README does not render the output documented:  see https://github.com/niceume/datasailr#code-example

Consider using the same example as the one you used in the [paper.md](https://github.com/niceume/datasailr/blob/joss_submission/paper.md)
