### 2021-04-19 niceume/datasailr#13 @crvernon
JOSS Review [crvernon]: Include masked regex behavior in documentation
**Review reference:** openjournals/joss-reviews#3166

It should be noted in the documentation that a duplicated value in a regex query string will use the first assignment in the subsequent conditional statement without notifying the user that a multiple value exists.  See the `Hornet` value that I added in for both the `usa` and `japan` assignments:

```
germany = re/(^Merc|^Porsche|^Volvo)/
usa = re/(^Hornet|^Cadillac|^Lincoln|^Chrysler|^Dodge|^AMC|^Camaro|^Chevrolet|^Pontiac|^Ford)/
japan = re/(^Mazda|^Datsun|^Honda|^Toyota|^Hornet)/
```

This will assign `Hornet` to `usa` when using your following conditional:

```
if ( _rowname_ =~ germany ) { country = "Germany" ; type = rexp_matched(1); }
else if( _rowname_ =~ usa ) { country = "USA"  ; type = rexp_matched(1);  }
else if( _rowname_ =~ japan ) { country = "Japan"  ; type = rexp_matched(1); }
else { country = "Other" }
```

This is expected when only considering the conditional statement; however, it could introduce unexpected behavior if the regex queries are incorrectly constructed (e.g., if `Hornet` has been incorrectly assigned to `usa` in the above).
