### 2021-04-20 fd78d21
example code and its output are updated.
