### 2021-04-20 niceume/datasailr#13 @niceume
Thank you for your comment!

Even when ^Hornet  appears in regular expressions for both usa and japan, they both work. (i.e. they do not mask each other.)

The following code shows that ^Hornet in usa and ^Hornet in japan work.

```
library(datasailr)
data(mtcars)

sail(mtcars, '
germany = re/(^Merc|^Porsche|^Volvo)/
usa = re/(^Hornet|^Cadillac|^Lincoln|^Chrysler|^Dodge|^AMC|^Camaro|^Chevrolet|^Pontiac|^Ford)/
japan = re/(^Mazda|^Datsun|^Honda|^Toyota|^Hornet)/

_rowname_ =~ germany
type_germany = rexp_matched(1)
_rowname_ =~ usa
type_usa = rexp_matched(1)
_rowname_ =~ japan
type_japan = rexp_matched(1)
')
```
