### 2021-04-20 niceume/datasailr#13 @crvernon
> Even when ^Hornet appears in regular expressions for both usa and japan, they both work. (i.e. they do not mask each other.)
> The following code shows that ^Hornet in usa and ^Hornet in japan work.

I agree that this situation works; however, when the assignment occurs for a common field (e.g., `type`) only the first occurrence of the value (`Hornet`) will be recognized.
