### 2021-04-20 niceume/datasailr#13 @niceume
> when the assignment occurs for a common field (e.g., type) only the first occurrence of the value (Hornet) will be recognized.

This is true when regular expression matching is conducted in if-else's conditions. (you already seem to have mentioned)

In the following code, the last regular expression matching result overwrites the previous results, because those matching results are assigned to the same field (`type`).

```
library(datasailr)
data(mtcars)

sail(mtcars, '
germany = re/(^Merc|^Porsche|^Volvo)/
usa = re/(^Hornet|^Cadillac|^Lincoln|^Chrysler|^Dodge|^AMC|^Camaro|^Chevrolet|^Pontiac|^Ford)/
japan = re/(^Mazda|^Datsun|^Honda|^Toyota|^Hornet)/

_rowname_ =~ germany
type = rexp_matched(1)
_rowname_ =~ usa
type = rexp_matched(1)
_rowname_ =~ japan
type = rexp_matched(1)
')
````

I am sorry if I misunderstand what you are saying.

What I am saying is that whether the first occurrence or the last occurrence takes effect does not come from regular expression's feature. It comes from how the code is executed.
