### 2021-04-20 niceume/datasailr#13 @crvernon
> What I am saying is that whether the first occurrence or the last occurrence takes effect does not come from regular expression's feature. It comes from how the code is executed.

I agree.  I didn't want users to be misled by assumptions that there would be some error raised if there were errors within their regular expressions, in terms of duplicated values in the query string.  Since these query strings will likely be constructed manually, I believe they are prone to these types user error.

If you believe that the current representation of how this functions is sufficiently represented in your documentation, I will close the issue.
