### 2021-04-21 30bea3f
Ver.0.8.8 corresponds to the work published on JOSS
