### 2021-04-21 niceume/datasailr#13 @niceume
I understand. When I make tutorials, I will keep it in mind. Thank you very much for your comments!
