### 2021-04-21 #3166 @DrMattG
> @fboehm I believe that @niceume has addressed all of the issues I submitted in the `joss_submission`branch of the `datasailr` repository here: https://github.com/niceume/datasailr/tree/joss_submission
>
> I have checked all of my review boxes and consider this submission as ready awaiting the approval of you and @DrMattG
>
> Excellent work @niceume !

I agree - great work @niceume. I am happy that the issues I raised are now addressed. Thanks @fboehm for the opportunity to review - it's a really nice process.
