### 2021-04-24 #3166 @niceume
@fboehm

* I have checked the generated wheedon PDF proof
* I created a release on GitHub for this version ( https://github.com/niceume/datasailr/releases/tag/joss0.8.8 )
* I have created a Zenodo-archive of this release.

The Zenodo doi is 10.5281/zenodo.4716943

Thank you.
