### 2021-04-24 #3166 @fboehm
@niceume - before we proceed, a few minor fixes are needed for the paper.

line 15 and other lines: row-wise (or rowwise - I believe that both spellings are acceptable)
line 20: domain-specific
throughout: insert spaces before parentheses around citations. So it should look like "we used R (R foundation, 2021)", where there's a space between R and (.
line 37: Do you want "datasailr" as the first word of the sentence?
line 70: Do you want "dplyr" instead of "Dplyr"? You can add curly braces around dplyr, like "{dplyr}", in the bib file to force rendering of lower-case letters.
line 66: same comment - do you want "sqldf"? and "R" is needed in place of "r".

Please make these fixes and double-check the resulting pdf.
