+++ paper.bib
@@ -7,7 +7,7 @@
 }
 
 @book{garcia2015,
-  title     = {Data Preprocessing in Data Mining},
+  title     = {{Data Preprocessing in Data Mining}},
   author = {Salvador García and Julián Luengo and Francisco Herrera},
   year      = {2015},
   publisher = {Springer International Publishing},
@@ -15,7 +15,7 @@
 }
 
 @Manual{dplyr:2021,
-  title = {dplyr: A Grammar of Data Manipulation},
+  title = {{dplyr: A Grammar of Data Manipulation}},
   author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller},
   year = {2021},
   note = {R package version 1.0.4},
@@ -23,7 +23,7 @@
 }
 
 @Manual{sqldf:2017,
-  title = {sqldf: Manipulate R Data Frames Using SQL},
+  title = {{sqldf: Manipulate R Data Frames Using SQL}},
   author = {G. Grothendieck},
   year = {2017},
   note = {R package version 0.4-11},
@@ -31,7 +31,7 @@
 }
 
 @Manual{magrittr:2020,
-  title = {magrittr: A Forward-Pipe Operator for R},
+  title = {{magrittr: A Forward-Pipe Operator for R}},
   author = {Stefan Milton Bache and Hadley Wickham},
   year = {2020},
   note = {R package version 2.0.1},
@@ -39,7 +39,7 @@
 }
 
 @book{sas1985sas:1985,
-  title={SAS user's guide: Statistics},
+  title={{SAS user's guide: Statistics}},
   author={{SAS Institute}},
   volume={2},
   year={1985},
