### 2021-04-26 #3166 @fboehm
Never mind - I found the version number in the archive.
