### 2021-05-01 #3166 @Kevin-Mattheus-Moerman
@niceume I have proodread you paper and have some minor comments to address before we can move on. Often the word `a` or `the` seemed missing from a sentence. Please implement these changes or respond if you disagree. Thanks.

- [ ] Can you rephrase the sentence `I present yet another way to process data in a row by row manner using data manipulation oriented script, DataSailr script.`, first of all please remove all occurrences of "I" (or "we" etc if any) and rephrase. Furthermore I think the word "a" should be added, for instance I propose: `This work presents another way to process data in a row by row manner using a data manipulation oriented script, DataSailr script.`
- [ ] Fix `Row-wise approach does not require vectors,....`, which should likely read `The row-wise approach does not require vectors,....`
- [ ] Fix `DataSailr script is a domain-specific language...`, which should likely read `The DataSailr script is a domain-specific language...`
- [ ] Fix `...the result vector to BMI column`, which should likely read `...the result vector to a BMI column`.
- [ ] Fix `... a main function called sail() function`, which should likely read `... a main function called sail()`.
- [ ] Fix `The first argument is data frame.`, which should likely read `The first argument is a data frame.`
- [ ] Fix `The second argument is DataSalr script,`, which should likely read `The second argument is the DataSalr script,`
- [ ] Fix `explained in README and DataSailr website`, which should likely read `explained in the README and on the DataSailr website`

Suggestion
- [ ] I notice your hearth felt and kind personal acknowledgement. This is the first time I've seen such a personal acknowledgement. Although I can identify with the sentiment, I do kindly ask if you could consider to only acknowledge professional/academic support. This can include financial support, technical support, advise, or the acknowledgement of contributors who are not authors on the paper.
