+++ paper.md
@@ -18,30 +18,30 @@ bibliography: paper.bib
 
 # Summary
 
-Data processing and data cleaning are essential steps before applying statistical or machine learning procedures [@pyle1999; @garcia2015]. R provides a flexible way for data processing using vectors. Additional R packages also provide other ways for manipulating data such as using SQL and using chained functions. I present yet another way to process data in a row by row manner using data manipulation oriented script, DataSailr script. The script enables users to instruct data processing in a row-wise manner, and the script is intuitive and easy to learn. This article introduces the datasailr package, and potential benefits of this row-wise approach.
+Data processing and data cleaning are essential steps before applying statistical or machine learning procedures [@pyle1999; @garcia2015]. R provides a flexible way for data processing using vectors. Additional R packages also provide other ways for manipulating data such as using SQL and using chained functions. This work presents yet another way to process data in a row by row manner using a data manipulation oriented script, DataSailr script. The script enables users to instruct data processing in a row-wise manner, and the script is intuitive and easy to learn. This article introduces the datasailr package, and potential benefits of this row-wise approach.
 
 
 # Statement of Need
 
-The [datasailr](https://CRAN.R-project.org/package=datasailr) package enables users to instruct data processing in a row-wise manner using DataSailr script, which is an easy and intuitive script designed especially for data manipulation. This row by row approach is different from R's built-in functionality, which provides data processing through column vector calculations. Row-wise approach does not require vectors, and can be intuitive compared to R's default column vector approach.
+The [datasailr](https://CRAN.R-project.org/package=datasailr) package enables users to instruct data processing in a row-wise manner using DataSailr script, which is an easy and intuitive script designed especially for data manipulation. This row by row approach is different from R's built-in functionality, which provides data processing through column vector calculations. The row-wise approach does not require vectors, and can be intuitive compared to R's default column vector approach.
 
-DataSailr script is a domain-specific language for data processing. Using domain-specific language for data manipulation is a similar approach to SAS^[SAS and all other SAS Institute Inc. product or service names are registered trademarks or trademarks of SAS Institute Inc. in the USA and other countries.] software [@sas1985sas:1985]. SAS software provides DATA blocks, within which users can write scripts that are specific for data manipulation in a row-wise manner. The separation of data manipulation steps and statistical procedures has a benefit of improved readability. Its row-wise data manipulation script is easy to understand and learn. DataSailr brings the same kind of experience to R.
+The DataSailr script is a domain-specific language for data processing. Using domain-specific language for data manipulation is a similar approach to SAS^[SAS and all other SAS Institute Inc. product or service names are registered trademarks or trademarks of SAS Institute Inc. in the USA and other countries.] software [@sas1985sas:1985]. SAS software provides DATA blocks, within which users can write scripts that are specific for data manipulation in a row-wise manner. The separation of data manipulation steps and statistical procedures has a benefit of improved readability. Its row-wise data manipulation script is easy to understand and learn. DataSailr brings the same kind of experience to R.
 
-For example, you have student health data and you need to calculate each student's body mass index (BMI) from their body weight and height. Using R's built-in functionality, you need to access body weight and height as column vectors, conduct vector calculation, and assign the result vector to BMI column. DataSailr script, on the other hand, enables you to write code as how to calculate BMI for each person.
[…]
