### 2021-05-02 #3166 @niceume
@Kevin-Mattheus-Moerman - Thank you very much for your reading and comments. I have fixed the parts you mention. I also understand that this is an academic/professional work, and the personal acknowledgement part is now removed from the paper.

@fboehm - If this looks ok, I will make a new release, archive it to Zonodo and obtain a new DOI.
