### 2021-05-02 #3166 @Kevin-Mattheus-Moerman
@niceume thanks for implementing those changes. I checked them and I think the paper looks good now. Since these are only editorial changes, from our perspective you do not need to create an updated release and archived version. Therefore if you wish I can process acceptance now.

If you do want the paper in the archived version to be up to date then feel free to do as you proposed. Let us know if the version tag changes too. I'll then update the archive link and the version tag and process acceptance at that point.
