### 2021-05-02 #3166 @niceume
@Kevin-Mattheus-Moerman - Thank you very much for having things proceed.

The release tag and DOI are updated as follows.
The release tag is joss0.8.8-3
The Zonodo DOI is 10.5281/zenodo.4732788
