### 2021-05-02 #3166 @Kevin-Mattheus-Moerman
Congratulations @niceume your paper has now been published in JOSS!

Thank you @fboehm for editing this work, and thank you very much @DrMattG and @crvernon for your review efforts!
